CORE-8074 ThreadLooper thread safety fixed (#2617)
The thread is started and the var set in a single line. This meant it was possible to enter and close the thread before thread was ever set in the instigating thread.

That is one of the reasons a lock exists, but the check defensive code was outside this lock. This was first caught in the code when we tried to stop the loop, which first checks that you are only trying that from within the loop thread, except thread was still null so it was assumed erroneously by this check that it was being called from outside the loop. That resulted in a throw (from the check(){}) which was caught only by the top level handler, thus _isRunning was never set to false. As a safety "fix" _isRunning is also now set, and the lifecycle ERRORed if the loop exists with an unexpected exception because that is technically the case on both parts.

All the vars which need to be set under lock are now moved to an inner class along with the lock. This should make the class a little more future proof/maintainable because it's easier to review what should be under the lock.