CORE-6624 - Review lifecycle handling for MemberOpsService (#2633)
Made changes to MemberOpsService's lifecycle handling to follow the recommended order for closing down and creating the resources.

Had to modify LifecycleTest.kt, because it was using the testClass's type as coordinator name which in our case would have required to rename the lifecycle coordinator to "MemberOpsServiceImpl" instead of being able to use the interface's name. I followed the protocol of the other functions that are using reified D as a type parameter. This way we can easily provide MemberOpsService as the type and it will not use the implementation class to retrieve it from the registry.