CORE-8622 Recreate transactions with empty component groups (#2646)
Empty component groups are not recorded when a transaction is persisted
to the database.

When loading a transaction from the database, it finds all the component
groups with the matching transaction id. These component groups are then
used to recreate a `WireTransaction`.

The issue is that the empty component groups which were not found in the
database, because we never recorded their existence affects the
calculated transaction id in `WireTransaction`. Causing the loaded
transaction to be subtly different and also have a different transaction
id than it was saved with.

To fix this, the number of component groups in the transaction is
recorded in the `TransactionMetadata` when initially creating a
`WireTransaction`. The metadata is used when recreating
`WireTransaction`s from the database, where we now know which groups
should exist and can therefore create an empty component group if the
database did not return anything for a group that should exist.