CORE-8099: Simulator ConsensualStatesLedger persistence
Persistence uses various JPA entities which have been explicitly added to the JPA persistence info.
Includes some tidy-up of node and flow set-up for better thread safety.
Simulator's keys are now generated using Corda's CipherSchemeMetadata, which was already being used to serialize them.
The schema for persistence is located in simulator.changelog-master.xml which is available in the API, consistent with Corda's approach. Note that all entities though are internal to Simulator runtime.