CORE-6800: Simulator accepts MemberX500Name, creates HoldingIdentity
Real Corda uses a GroupPolicy file with a group identity together with membership to create its nodes. In Simulator however each instance of Simulator is effectively a stand-alone group. Rather than creating nodes using a HoldingIdentity, we revert to using just MemberX500Name, and create the HoldingIdentity as part of the nodes, passing in the GroupId for that instance of Simulator.