CORE-9009 Improve entity manager factory handling for virtual node reconcilers (#2781)
This change alters the behaviour of the reconciler context handling so that they are managed as streams of contexts which do not initialise their entity managers until necessary. This should ensure that we are only creating database connections when necessary instead of creating all up front which will reduce the number of concurrent connections when there is a large number of virtual nodes.