CORE-7918: Re-poll messages when repartitioned in the middle of processing (#2813)
* CORE-7918: Re-poll messages when repartitioned in the middle of processing.


We reset the poll interval while processing batches of events. This leads to a situation where we have a batch, call the poll, and end up with a repartition event which resets our message offsets to the messages we've just pulled off the bus. In these cases the simplest thing is to just drop the messages we're processing and poll for them again. Otherwise we'll end up reading the same message twice.

This change will have the rebalance listener set a flag after a rebalance so that we can check for it and abandon the messages.