CORE-9065 - Add pre-auth token and skip-approval mode for corda-cli (#2863)
Include optional parameter for the onboard member plugin to accept a pre-auth token and include it in the registration context sent to the MGM.
e.g. --pre-auth=
Existing behaviour for waiting for approved registration can be used if flag is set
e.g. --wait-for-approval
If flag is not set, return registration ID

Also added a fix for broken cluster deployment.