CORE-8984 Add extra logging to finality and notary flows
Add extra `trace` and `debug` logging to the finality and notary flows.
These logs have been added to identify an issue in the UTXO finality
flows; however, these logs will generally be useful when identifying
issues in production when more detail is needed.

Changes to the log4j2 files have been made and should be reverted when
the underlying issue has been diagnosed and fixed.