CORE-9166: Cancel Event Processing & Resume Partitions on Rebalance (#2939)
The state and event message pattern internally pauses the eventConsumer
and polls from it while resetting the polling interval and waiting for
the attached processor to finish, mainly to prevent the consumer from
being kicked out of the group due to the poll interval timeout being
reached. If a rebalance (caused by another consumer joining/leaving the
group) occurs while executing the mentioned polls, the state and event
message pattern might end up resuming incorrect partitions, which can
lead to either processing events more than once or processing events
for which the associated state is not yet in sync. When dealing with
the flow-worker, these scenarios lead to flows incorrectly being sent
to DLQ and stuck in unexpected states.

Changes within this commit attempt to prevent the above by keeping
better tracking of the dirty flag set upon rebalance and aborting
events mid flight whenever necessary.