CORE-9471 - Fix issue with views (#2956)
Description of the issue:
Unable to get registration status of member’s registration ID using the MGM’s holding identity on the MGM’s cluster.

Fix:
The built query contained a filter for the holding identity by mistake. The holding ID provided for querying should only determine which member's view we want to see. Not whose request we want to see.