CORE-9807: Fix Status Transition in RPCSubscription (#3013)
The listener sets the subscription status as 'DOWN' whenever there are
no partitions assigned and back to 'UP' again within
'onPartitionsAssigned' callback. The latter, however, is always invoked
when the cooperative rebalance algorithm is used (it might receive an
empty list of partitions or not), so it should set the status as 'UP'
if and only if the current set of assigned partitions is empty and the
received list of partitions is not, otherwise it might wrongly
transition the subscription to 'UP' even though it should have stayed
as 'DOWN'.