CORE-9451: rework crypto config (#3025)
Previously the crypto code was tightly coupled with the config service. In particular:

1. The crypto service created its own default configuration, with embedded passphrase, and then for the combined worker stored this in its database. (Everything else in Corda uses the Config schema to handle defaults). Special measures were taken to make sure the default master wrapping key passphrase was encrypted, which required knowledge of how smart config handles secrets which is no longer safe to assume since we allow different secrets implementations.

2. Different crypto service providers were supported, each with its own configuration. In order to handle this in a general way, the smart configuration for the crypto service was converted to JSON by CryptoServiceFactoryImpl, then dynamically instantiated into configuration objects particular to a specific crypto service provider, and passed in dynamically typed. This also involved care to make sure that secrets such as the passphrase could still be decrypted, which make assumptions that secrets are always encrypted locally. With other sources of secrets that assumption is no longer sound.

3. The configuration of the crypto service providers was handled by a layer of data classes allowing plain access to the classes, rather than using methods of SmartConfig. That's reasonable of itself, but is different to the rest of Corda.

4. Some of the crypto code at lib level was configurable, which would normally have caused a dependency on the confgiuration library. However, the configuration library has a dependency on the crypto library, in order to handle cryptogrpahy on the the secret parts of the config. This prevented some parts of the crypto code from using SmartConfig.

5. There was a separate Corda crypto technical compliance kit which shares code with the production implementation. In combination, this means that the configuratioon of crpto scheme is complex and brittle, and would get significantly more complex when extended to include support for third party secret stores.

In this change, I've moved to mostly using SmartConfig directly, much like the rest of Corda. I've also moved the library-level generic (templated) crypto service provider, and instead we have the components/crypto/crypto-softhsm-impl provide a simplified, non-templated/non-generic crypto service provide single abstract method interface and a simple implementation which provides a simplified type for the crypto service. We may well never need the flexbility of different config service providers, and it is expensive to maintain and would have made this change impractical.

Having components/crypto/crypto-softhsm-impl define the CryptoServiceProvider isn't ideal, but:
a) it fits there without introducing a dependency cycle with crypto-service-impl
b) we plan to shortly collapse down the support for multiple different types of HSM, and merge crypto-softhsm-impl and crypto-sevice-impl

The Crypto service at this point still has its own way of constructing default configuration to go into the database. We may get rid of this soon, and instead simply use the JSON config schema as the rest of the Corda does. For now we pass in the passphrae and salt default values.

Extend some comments and a README on how config works based on what I figured out about configuration to design and implement this change.

Add a toString() to dump out AesKeys in Base64. To be used with care, but if you need to actually compare keys to debug this helps.

Delete KeyCredentials, its tests and methods in other classes with initialise based on it, since we no longer use it. I simply
have the createDefaultCryptoConfig take two parameters instead of needing a dataclass.