CORE-9062 Add rule type to the group approval rules delete message (#828)
Add rule type to the delete avro schema for approval rules so that we can segment rules in the database between standard and preauth. This is mainly so that we can have separate delete endpoints for standard and preauth rules and avoid a scenario where we can delete preauth rules through that standard rules endpoint, and vice versa.