CORE-10050: remove TransientSoftWrappingKeyMap and the config parameter which selects between it and CachingSoftWrappingKeyMap
CachingSoftWrappingKeyMap can do everything TransientSoftWrappingKeyMap can do, and if you configure it to zero size the behaviour is identical. There is a case for having a version without caching selectable, but that has to be weighed against the extra complexity. Ultimately, simplicity or least a level of comprehensibility is important to make the implementation secure, and we need to move toward that.