CORE-9064 Implementation of pre-auth tokens rules logic (#3111)
Changes required to manually review registrations with a pre-auth token.

This includes a fix so that registration requests can be declined even if the pending member info has not yet been persisted. Previously registration requests could not be moved to declined state if the member info had not yet been persisted.
Pre-auth tokens are consumed if present, valid, and pre-auth token rules can be queried.
Pre-auth tokens are validated on initial request verification, and again before being consumed. Registrations are declined if the pre-auth token provided is invalid.