CORE-9717 Vnode platform migration (#3152)
Introduces a new plugin sub command in the vnode plugin which migrates virtual nodes to the latest version of the platform schema. Handles vault, crypto and uniq schemas.

The command can be executed along the lines of:

corda-cli vnode platform-migration --jdbc-url=jdbc:postgresql://localhost:5432/cordacluster -u user

This will read a file holdingIds which must contain a list of short holding ids of virtual nodes to migrate. The db access required is read only for liquibase tracking purposes. It will output the necessary sql commands into a file vnodes.sql which can then be applied to the db. Both the input and output file paths are configurable from the command line too, as is the db password (omitted in the example above because the cluster db created by Helm doesn't require one).