CORE-9342 Allow corda kafka consumers to reassemble chunks into compete records which were created by the corda kafka producer (#3040)
Update the CordaKafkaConsumer to handle chunked records which were generated by the CordaKafkaProducer.
Should ignore chunks if the consumer is expecting to read records of type Chunk. i.e sent via app level chunking.
Deserialization errors and checksum errors are handled within the ConsumerChunkDeserializerService which is built using the same onError handler as the deserializer.

Chunked records generated by the CordaKafkaProducer are always sent via transactions so ordering is always preserved with a single transactions. Therefore the final chunk always has the checksum and indicates when we can try to reassemble the chunk.

Chunks may be interleaved with chunks from a different transaction or normal records. We should not return any records from a partition which has partial chunks being read. Given chunks are always committed via transactions it should be impossible for a partition to be blocked by incomplete chunks. records are buffered in memory within the consumer per partition. This means if there is a rebalance the same records can be reread and no records will be lost. If our assignment changes after a rebalance we should remove any records buffered for partitions lost to ensure we do not have stale records if we are reassigned those partitions