CORE-10466: Remove call to `lb.close()`
As it is tinkering with `java.sql.Connection` which was passed into the method and which we did not create/own and therefore should not be closing.

Without this change test `net.cordacon.example.rollcall.TruancyResponderFlowTest.should verify records sent against the key and persist them` will fail as we corrupt the connection inside `net.corda.simulator.runtime.persistence.DbPersistenceService`.