CORE-8896 Extract Helm logic to helpers (#3266)
Extract all of the Helm chart logic to helpers as a prelude to moving it
into a library chart for reuse. This is almost entirely internal
refactoring. Merging the REST and P2P Gateway service implementations
resulted in two external changes. The port for the REST service is now
configurable (still defaults to 443) and the P2P Gateway service is no
longer headless (`clusterIP: None` is no longer set).