CORE-10409 Generate chunk cleanup records for old states (#3210)
When updating a state record in the state and event pattern we should remove any ChunkKeys for the state that are no longer valid. Kafka compaction will remove keys which have a newer value so we only need to cleanup keys for chunked states that get smaller or set to null.

Removed requestId from ChunkKey to make ChunkKeys deterministic. This means if the chunks get smaller we know we can safely remove the ChunkKeys for the higher partNumbers no longer applicable