CORE-8444 do not log secrets as part of config validation (#3229)
Updated the boot config schema to remove values not used anymore related to message pattern db bus configs and to populate the schema for the latest db values so they can be validated properly.
Updated the validator to log the configs being validated with secrets hidden. moved config logging to debug as validation should not be at info.
Added logic to make the db password a secret in the boot config.
fixed an issue where we were always setting the bus type to be KAFKA in the kafka section of the messaging config in the boot config, even if it was the DB bus. This was masked by an override in the db message bus merger.