CORE-5533 - minimum config value setting not being adhered to when config change is input (#905)
Description:
The issue here: CORE-5533 seemed like the minimum setting of 1000 milliseconds for maxRetryDelay was not being used as 999 milliseconds was accepted.

However, looking at the query that was tested actually the issue appeared to be that the maxRetryDelay was not queried within the processing object it was defined in. As a result of this the maxRetryDelay was actually being set and accepted as an additional property.

The fix here was to prevent additional properties from being accepted. This was done by adding additionalProperties: false to: properties object, event object, session object, processing object.

The issue raised was replicated and then the solution tested using manual testing.