CORE-8744 Align Cpi related entities and repository with latest JPA guidance
File checksums and cpi signer summary hashes are now strongly typed. An instance
of SecureHash is now used instead of String since the former provides more context to people.

CpiIdentifier type should be used in the methods instead of separate arguments.
This helps keeping the list of arguments for a method small.

Deleted PersistenceUtils.kt to simplify the code base. To help reduce complexity of the code,
variables should always have the same name across the code base and not variations of the
original name. If someone requires a string version of a SecureHash, they can call the method toString().

Other but small refactors were also made.