CORE-9865 Creation of InterOpFacade Processor - sending additional seed messages
Added 2 new seed messages to trigger search for and usage of InteropGroupPolicy 1) the hardcoded interop policy if found and the message is sent to gateway, 2) the group policy id doesn't match and message is dropped. Imporved logs in LinkManager- log only 1 statement for a decision about an incmping message, a duplicated entry (start) "Processing outbound" removed.