CORE-10678 Make subsequent MGM Registrations Fail (#3258)
Make subsequent MGM Registration's fail. When registering the MGM check if there is already a registered MGM for a given holding identity and fail if it has status approved. Prevent concurrent registrations for the same holding identity by keying the registrations request topic by holding id short hash.

To help with robustness. I have also made the database operations: PersistGroupPolicy and PersistMemberInfoHandler idempotent (PersistGroupParametersInitialSnapshot was made idempotent as apart of CORE-10347). I have also MGM Registration so that we persist the registration request with status approved as the last database operation. Previously we were doing that at the beginning, this means that MGMRegistration will now be retried, if database operations fail (see CORE-9972).