CORE-10286, CORE-10394, CORE-10277: collapse a few classes into SoftCryptoService (#3164)
Have SoftCryptoService take a wrapping key store object and 2 caches, and the root wrapping key.

Eliminate the config classes that encapsulated the configuration of the crypto service layer. The only dynamic configuration was for the caches, which are now passed in. It is therefore possible to simply configure the caches in SoftCrytpoServiceProviderImpl instead.

Port all tests from the deleted classes, removing a few duplicates. Rework some of the test cases to avoid test instrumentation. Clean up the test code to reduce repetition. Share test fixtures where possible, which improves test runtime.

Inject key pair generator and wrapping key factory to the SoftCryptoService to support mocking out or replacing for testing.

Make WrappingKey an interface, so that it can be implemented for testing.

Move out the long list of acceptable key and signatures combinations into its own file.

Reduce log levels.

Rename WrappingKeyAlias.masterWrappingKey to wrappingKeyAlias for clarity.