CORE-6044: change health port for combined worker
Use port 7004 for combined worker since it is often run on Macs where port 7000 is not available. At this tage we won't change the separated workers since that may require significant downstream changes and separated workers are typically run in containers where there is unlikely to already something is already listening on 7000.