CORE-5630 - Add support for CPK metadata files
Added the feature to load configuration files for external channels for each cpk.
It is assumed the configuration format is JSON but no validation has been implemented yet.
CpkLoaderV2 is responsible for loading the file containing the configuration for external channels.
The class CpiValidatorImpl is responsible for validating the JSON. (Keep in mind that the logic for validating the configuration is not implemented, yet).
The configuration for external channels is stored in the CpkMetadata.
Added unit-tests.
Updated other tests that are related to the CpkMetadata.