CORE-11034: Add a cache per sandbox type and make the size of each configurable (#3344)
The cache is stored away in MultiCache which holds a map of one cache per SandboxGroupType. Where possible, the calling interfaces weren't changed. In some cases (see CacheControl, for example) the SandboxGroupType had to be added to distinguish between which cache a call was for. This should be external to any Sandbox users, however, as at their level it's the VirtualNodeContext which they'll (still) be passing in.