CORE-11640: start JPA repository layer, currently inert (#3346)
We should follow the JPA repository pattern in the Crypto persistence code. In this change we add the new code, not referenced yet from production code. Take care to make the new structure highly unit testable, by developing it using test driven development.

In future releases, we will need to support upgrade of the Crypto schema, which is challenging since tables are shared between cluster and virtual nodes. In production the crypto processor has many database connections, most to per-virtual-node databases and a small number to cluster databases. The timing of virtual node upgrade is up to the owner of the virtual node; we will not require everyone to upgrade at the same time. Therefore, structure the code with a "V1" in the schema name, so that we can later have "V2", "V3" etc, and potentially use any of them on a per database choice.

Stop caching the virtual node database connections. DbConnectionManager is the right place to do that, though it doesn't do so yet.

Take care over closing down the EntityManager and EntityManagerFactory when appropriate, but not for the EntityManagerFactory for the cluste which is owned by DbConnectionManager.