CORE-9595 System flow versioning (#3238)
Provides the ability to version inline system flows so that a group of 
sessions can guarantee that they are executing the same version of a 
flow pair.

System flows are versioned by platform version.

A system flow may or may not provide new versions of the flow. When a 
new version is added (versioning must be used at the start or a breaking
change must occur) that version will be used if the platform version 
agreed amongst sessions is high enough.

The platform version of a flow pair is determined by finding the lowest 
common platform version amongst a group of sessions. This platform 
version information is extracted from the MGM via the `MemberLookup` 
service.

To version a flow call:
- `VersioningService.versionedSubFlow` with a 
  `VersionedSendFlowFactory` on the initiating/starting side of the 
  flow pair.
- `VersioningService.versionedSubFlow` with a 
  `VersionedReceiveFlowFactory` on the receiving side of the flow pair.
  
The versioned flow factories construct instances of versioned flows. 
They receive a version and then the factory should have a `when` block 
that returns a `SubFlow` instance that supports that platform version.

The versioning flows are not interacted with directly. These flows 
manage the sending and receiving of version information. This includes 
the creation of `FlowSession`s that piggyback a session's first within a
versioned flow to also send version information. There is then a 
matching receiving session to pair with this logic.

When a platform version has been determined (on both sides), the version
is placed into the flow context properties. If a new versioned subFlow 
is executed when there is already a version in the context properties, 
then that version will be used and the sending of the version will be 
skipped. This allows us to simplify the code.
 
The currently versioned flows are:
- `ConsensualFinalityFlow`
- `ConesnsualReceiveFinalityFlow`
- `UtxoFinalityFlow`
- `UtxoReceiveFinalityFlow`
- `TransactionBackchainResolutionFlow`
- `TransactionBackchainSenderFlow`