CORE-11521 Remove `DigitalSignature.WithKey.context` (#929)
`DigitalSignature.WithKey.context` is returned as part of the signature. Callers of sign operation have no control over it so it's of no use to them. As such `DigitalSignature.WithKey.context` cannot be signed and is not signed. `DigitalSignature.WithKey.context` was originally added to carry untrusted (unsigned) metadata but was not really used.

* Removes `DigitalSignature.WithKey.context`