CORE-11065 Custom ledger queries liquibase changes (#941)
Add `custom_representation` and repurpose the `consumed` column
in the `utxo_relevant_transaction_state` table.

The `custom_representation` column is of type `JSONB` and is
therefore Postgres  specific.

`ledger-utxo-relevant-transaction-state-creation-v1.0.xml` has been 
separated to create the table. For now it has been restricted to 
creating the table only for Postgres databases, meaning that the table
will not be made for anything else. In the future, we can add the table
with the correct columns when we support more databases.

Remove the original boolean held in the `consumed` column and store the
timestamp there instead. This is because we know a state is consumed if
this timestamp is filled in, so having a separate column was redundant.