CORE-6564 Add API to get a counterparties flow information such as protocol version (#3341)
Add a new KeyValuePairList to the SessionInit message to send session specific context properties from initiating to initiated parties. This will contain the protocol name and versions supported by the initiator.

Add a new SessionEvent payload type, SessionConfirm, which will be sent in response to a SessionInit message instead of the usual SessionAck. SessionConfirm will contain the initiated parties session specific properties. This will include the protocol version chosen based on the initiators supported versions using the protocolStore.

Session properties exchanged will be stored in the session state under counterpartySessionProperties. This will make it accessible to the Flow Engine.

Add API to the FlowSession, getCounterPartyFlowInfo(), which matches the old corda4 API to get the other party in the session's FlowInfo.

When the API is called and no data messages have yet to be exchanged then trigger session initiation. (Previously, session initiation was only triggered upon the first data send to avoid unnecessarily starting flows that were never interacted with.)
Wait until session initiation is complete before passing the FlowInfo back to the Flow Fiber.

If the session was already initiated then the flow will be given the FlowInfo without suspending