CORE-9586 Add Custom Fields To Metadata In MemberInfo (#3385)
Allow a user to specify custom fields in the registration context, which get signed and added to the member.info during registration. These fields must start with "ext.", everything else is reserved for future platform usage. Verify there are at most 100 custom fields being added and limit the key size and value size to 128 and 256 characters respectively. We verify this on both the member and MGM side during registration. Currently we don't verify any of the other fields on the MGM side, so we have future flexibility i.e. a cluster running a more recent version of Corda, can register with an MGM running an old version. See schema update here: corda/corda-api#936