CORE-11773: instantiate and pass around CryptoRepository (#3402)
Remove WrappingKeyStore - the interface and test coverage, since all that functionality is provided by CryptoRepository instead.

For simplicity, CryptoRepository is a simple function, which takes the references needed to construct itself as function parameters rather than OSGi dependency injection.

Rely on integration and smoke tests for covering lifecycle and dynamic config, dropping the duplicate alternative implementation in SoftCryptoServiceProviderImpl. This means we have integration and smoke test coverage but not unit test coverage of SoftCryptoServiceProviderImpl, which is a piece of glue logic to plug a couple of caches into SoftCryptoService and pass in some OSGi dependencies.