CORE-11709: Point HSMStore to use CryptoRepository (#3448)
* CORE-11709: Point HSMStore, SigningKeyStore to use CryptoRepository

Pointing those two classes to use the CryptoRepository means less code duplication and a path towards eventually removing them.  There's a lot of tests that need to be managed, however, in order to get that to work so this is a simpler and more expedient way to achieve that result.

Also ensuring that uses of CryptoRepository are closed to save database connections.