CORE-9596 Vault named query parsing (#3436)
Adds vault named query parsing of a Postgres syntax into an expression and back into a Postgres query to be executed at some point.

`VaultNamedQueryParser` provides the entry point for parsing a query. The implementation of this interface then delegates to a `VaultNamedQueryExpressionParser` to convert a query string into an expression, a `VaultNamedQueryExpressionValidator` to validate the expression and a `VaultNamedQueryConverter` that converts the expression into an output query string.

The `PostgresVaultNamedQueryExpressionParser` converts a Postgres query string into an expression, and `PostgresVaultNamedQueryConverter` converts that expression into a Postgres query string.

These interfaces give us the flexibility to parse from different formats and output queries for different databases.